refactor: moved sensor data source to `algorithms/libs` as well
